+++ examples/4qubit_gates.jl
@@ -1,7 +1,8 @@
 function decompose_CNot_41()
     # Reference: https://doi.org/10.1109/DSD.2018.00005
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
+        
     "num_qubits" => num_qubits,
     "maximum_depth" => 10,
     "elementary_gates" => ["H_1", "H_2", "H_3", "CNot_1_3", "CNot_4_3", "Identity"],
@@ -9,7 +10,6 @@ function decompose_CNot_41()
     "objective" => "minimize_depth",
     )
 
-    return params
 end
 
 function decompose_quantum_fulladder()
@@ -31,15 +31,15 @@ function decompose_quantum_fulladder()
         return CNot_2_4 * CV_4_2 * CVdagger_1_2 * CVdagger_3_2 * CNot_4_3 * CNot_3_1 * CV_1_2
     end
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
+
     "num_qubits" => num_qubits,
     "maximum_depth" => 7,
     "elementary_gates" => ["CV_1_2", "CV_4_2", "CV_3_2", "CVdagger_1_2", "CVdagger_4_2", "CVdagger_3_2", "CNot_3_1", "CNot_4_3", "CNot_2_4", "CNot_4_1", "Identity"],
     "target_gate" => target_gate_1(),
     "objective" => "minimize_depth",
     )
-    
-    return params
+
 end
 
 function decompose_double_toffoli()
@@ -58,15 +58,15 @@ function decompose_double_toffoli()
         return CV_2_4 * CNot_1_3 * CNot_3_2 * CV_3_4 * CVdagger_2_4 * CNot_3_2 * CNot_1_3
     end
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
+
     "num_qubits" => 4,
     "maximum_depth" => 7,
     "elementary_gates" => ["CV_1_2", "CV_2_4", "CV_3_4", "CVdagger_1_2", "CVdagger_2_4", "CVdagger_3_4", "CNot_1_3", "CNot_3_2", "CNot_2_3", "Identity"],
     "target_gate" => target_gate(),
     "objective" => "minimize_depth"
     )
-    
-    return params
+
 end
 
 function decompose_double_peres()
@@ -86,13 +86,13 @@ function decompose_double_peres()
         return CV_2_4 * CNot_1_2 * CV_3_4 * CV_1_4 * CNot_2_3 * CVdagger_3_4
     end
 
-    params = Dict{String, Any}(
+    return Dict{String, Any}(
+
     "num_qubits" => 4,
     "maximum_depth" => 7,
     "elementary_gates" => ["CV_1_4", "CV_2_4", "CV_3_4", "CVdagger_1_4", "CVdagger_2_4", "CVdagger_3_4", "CNot_1_2", "CNot_2_3", "Identity"],
     "target_gate" => target_gate(),
     "objective" => "minimize_depth"
     )
-    
-    return params
+
 end